refactor: change the structure of the folder
