+++ src/components/tables/data-table.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { Button } from "@/components/ui/button";
 import {
 	Table,
 	TableBody,
@@ -16,7 +15,7 @@ import {
 	getPaginationRowModel,
 	useReactTable,
 } from "@tanstack/react-table";
-import { DataTablePagination } from "./pagination";
+import { DataTablePagination } from "../tables/pagination";
 
 interface DataTableProps<TData, TValue> {
 	columns: ColumnDef<TData, TValue>[];
